chatServer/requirements.txt
asyncio==3.4.3
Django==2.2.6
django-cors-headers==3.1.1
django-logpipe==0.3.1
djangorestframework==3.10.3
kafka-python==1.4.7
lru-dict==1.1.6
pkg-resources==0.0.0
pytz==2019.3
sqlparse==0.3.0
websockets==8.0.2
